refactor: 모달이 한 번만 생기도록 TrendySection과 NewsletterCard 구조 개선
